Add files via upload
VISSIM-COM-trajectory-optimization
The current version of the script dynamically adjusts the speed of priority-equipped vehicles (connected vehicles) to optimize their passage through intersections during green signals. It calculates the time remaining until the next green or red phase and the distance from the vehicle to the signal head to compute an optimal speed.